Write underlying property name for inline class to metadata
This will be used in the compiler starting from 1.5.20 instead of the
currently used approach of looking for the single value parameter of the
primary constructor. The problem with the current approach is that
primary constructor can be private (since 1.4.30), and relying on
private declarations from metadata is dangerous; for example, until
recently, lazy IR didn't create stubs for private declarations at all,
which led to problems like KT-44723.